make it so more piano keys = lower range
Before, the first note of the piano was always the first note of the
octave. Meaning that the wider the piano, the higher the pitch. Now the
octave is "centered", meaning that on the 3-octave landscape piano, the
middle octave is the same as the base octave of the 1.5-octave portrait piano.