replace triangle and saw waves
The sawtooth wave has been replaced with a quarter sine. They have the
same spectral profile, but the phase relationships on the quarter sine
are very novel and more applicable to FM synthesis.

The triangle wave did not seem very useful, it's basically just a sine
wave with some weak harmonics on it, so it was removed.